add largest rectangle solution
This solution seems correct because it passes all spec unit tests.
But it encountered 'time limit exceeded' error when I submit it to
leetcode. Still need to find a better solution.

The tuned version is not correct, I wrote it after take a glance of
solution on book, but apparently I didn't get the point.

Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>